+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Change log
 Unreleased_
 -----------
 
+Changed
+~~~~~~~
+
+* Use lazy initialisation for HID enumeration on Windows so that the service
+  will not stop responding if there are a lot of devices.
+
 1.0.0_ |--| 2021-04-12
 ----------------------